requirements.txt
Pandas==2.0.1
numpy==1.24.3
plotly==5.14.1
PySimpleGUI==4.60.5
Pillow==9.5.0
nbformat>=4.2.0
matplotlib==3.7.1
wordcloud==1.9.2